feat(auth): configurable tenant id
This change allows the user to specify their tenant ID to the Keystone
endpoint. When authenticating against real Keystone this is unnecessary,
because it will determine the tenant ID from the username and API key.
However, it's useful for authenticating against Mimic when you need it
to work with other API clients. For instance, it's necessary for using
pandamonkey, Cloud Intelligence and Mimic all together.